+++ src/main/java/com/springbootlearning/learningspringboot3/ApiController.java
@@ -1,5 +1,6 @@
-package com.springbootlearning.learningspringboot3.ch2;
+package com.springbootlearning.learningspringboot3;
 
+import com.springbootlearning.learningspringboot3.ch2.Video;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
