+++ src/main/java/com/springbootlearning/learningspringboot3/Video.java
@@ -1,0 +1,4 @@
+package com.springbootlearning.learningspringboot3;
+
+record Video (String name) {
+}
