+++ src/main/java/com/springbootlearning/learningspringboot3/VideoSearch.java
@@ -1,4 +1,4 @@
-package com.springbootlearning.learningspringboot3.ch2;
+package com.springbootlearning.learningspringboot3;
 
 record VideoSearch(String name, String description) {
 }
