--- src/main/java/com/springbootlearning/learningspringboot3/ch2/Video.java
@@ -1,4 +1,0 @@
-package com.springbootlearning.learningspringboot3.ch2;
-
-record Video (String name) {
-}
